- CRUD de detalhes do plano - feito relacionamento entre as duas tabelas - Adicionado novas funcionalidades

+++ app/Http/Controllers/Admin/PlanController.php
@@ -5,7 +5,6 @@ namespace App\Http\Controllers\Admin;
 use App\Http\Controllers\Controller;
 use App\Http\Requests\StoreUpdatePlanRequest;
 use App\Models\Plan;
-use Illuminate\Support\Str;
 
 class PlanController extends Controller
 {
@@ -36,13 +35,7 @@ class PlanController extends Controller
 
     public function store(StoreUpdatePlanRequest $request)
     {
-        $data = $request->all();
-
-        if (isset($request->name)) {
-            $data['url'] = Str::kebab($request->name);
-        }
-
-        $this->repository->create($data);
+        $this->repository->create($request->all());
         return redirect()->route('plans.index');
     }
 
@@ -57,9 +50,18 @@ class PlanController extends Controller
 
     public function destroy($url)
     {
-        $plan = $this->repository->where('url', $url)->first();
+        $plan = $this->repository
+            ->with('details')
+            ->where('url', $url)->first();
+
         if (!$plan)
             return redirect()->back();
+
+        if ($plan->details->count() > 0) {
+            return redirect()
+                ->back()
+                ->with('error', 'Existem detalhes vinculados a esse plano, não e possivel deletar');
+        }
 
         $plan->delete();
         return redirect()->route('plans.index');
